requirements/common.txt
logzero
python-dotenv
discord.py
